Split ChatPanel send lifecycle tests
